feat(schemaExplorer): separate TimeRangeLabel
